services/drive-intel/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
scenedetect[opencv]==0.6.2
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
ultralytics==8.0.221
paddlepaddle==2.5.2
paddleocr==2.7.0.3
faster-whisper==0.10.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1
